refactor signed url logic into its own class
